requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
greenlet==2.0.2
    # via sqlalchemy
markdown-it-py==3.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
pygments==2.16.1
    # via rich
python-dotenv==1.0.0
    # via -r requirements.in
rich==13.5.2
    # via -r requirements.in
sqlalchemy==2.0.17
    # via -r requirements.in
typing-extensions==4.7.1
    # via sqlalchemy
